runtime detection fix

+++ lib/SmallChange/nodes/Scenery.cpp
@@ -79,12 +79,10 @@ Scenery::~Scenery()
   delete this->blocksensor;
   delete this->loadsensor;
 
-#ifdef HAVE_LIBSCENERY
-  if (this->system) {
+  if (sc_scenery_available() && this->system) {
     sc_ssglue_view_deallocate(this->system, this->viewid);
     sc_ssglue_system_close(this->system);
   }
-#endif // HAVE_LIBSCENERY
   delete this->pvertex;
   delete this->facedetail;
   cc_hash_apply(this->texhash, hash_clear, NULL);
@@ -97,9 +95,9 @@ Scenery::initClass(void)
   static int first = 1;
   if (first) {
     first = 0;
-#ifdef HAVE_LIBSCENERY
-    sc_ssglue_initialize();
-#endif // HAVE_LIBSCENERY
+    if ( sc_scenery_available() ) {
+      sc_ssglue_initialize();
+    }
     SO_NODE_INIT_CLASS(Scenery, SoShape, "Shape");
   }
 }
@@ -156,7 +154,7 @@ raypick_post_cb(void * closure)
 int 
 Scenery::render_pre_cb(void * closure, ss_render_pre_cb_info * info)
 {
-#ifdef HAVE_LIBSCENERY
+  if ( sc_scenery_available() ) {
   Scenery * thisp = (Scenery*) closure; 
   SoState * state = thisp->currstate;
   
@@ -232,15 +230,15 @@ Scenery::render_pre_cb(void * closure, ss_render_pre_cb_info * info)
     }
   }
   return 1;
-#else // !HAVE_LIBSCENERY
-  return 0;
-#endif // !HAVE_LIBSCENERY
+  } else {
+    return 0;
+  }
 }
 
 void 
 Scenery::GLRender(SoGLRenderAction * action)
 {
-#ifdef HAVE_LIBSCENERY
+  if ( sc_scenery_available() ) {
   if (this->system == NULL) return;
   if (!this->shouldGLRender(action)) return;
 
@@ -380,13 +378,13 @@ Scenery::GLRender(SoGLRenderAction * action)
     state->pop();
     SoGLLazyElement::getInstance(state)->reset(state, SoLazyElement::DIFFUSE_MASK);
   }
-#endif // HAVE_LIBSCENERY
+  }
 }
 
 void 
 Scenery::rayPick(SoRayPickAction * action)
 {
-#ifdef HAVE_LIBSCENERY
+  if ( sc_scenery_available() ) {
   sc_ssglue_view_set_culling_pre_callback(this->system, this->viewid,
                                    raypick_pre_cb, action);
   sc_ssglue_view_set_culling_post_callback(this->system, this->viewid,
@@ -398,13 +396,13 @@ Scenery::rayPick(SoRayPickAction * action)
                                    NULL, NULL);
   sc_ssglue_view_set_culling_post_callback(this->system, this->viewid,
                                     NULL, NULL);
-#endif // HAVE_LIBSCENERY
+  }
 }
 
 void 
 Scenery::callback(SoCallbackAction * action)
 {
-#ifdef HAVE_LIBSCENERY
+  if ( sc_scenery_available() ) {
   // FIXME: not correct to just evaluate in the callback()
   // method. SoCallbackAction can be executed for a a number of
   // reasons. Consider creating a SoSimlaEvaluateAction or something...
@@ -435,13 +433,13 @@ Scenery::callback(SoCallbackAction * action)
                                    NULL, NULL);
   sc_ssglue_view_set_culling_post_callback(this->system, this->viewid,
                                     NULL, NULL);
-#endif // HAVE_LIBSCENERY
+  }
 }
 
 void 
 Scenery::generatePrimitives(SoAction * action)
 {
-#ifdef HAVE_LIBSCENERY
+  if ( sc_scenery_available() ) {
   if (this->system == NULL) return;
   SoState * state = action->getState();
 
@@ -457,7 +455,7 @@ Scenery::generatePrimitives(SoAction * action)
   this->pvertex->setDetail(&pointDetail);
   this->curraction = action;
   sc_ssglue_view_render(this->system, this->viewid);
-#endif // HAVE_LIBSCENERY
+  }
 }
 
 void 
@@ -483,7 +481,6 @@ Scenery::loadsensor_cb(void * data, SoSensor * sensor)
   thisp->setLoadRottger(thisp->loadRottger.getValue());
 }
 
-#ifdef HAVE_LIBSCENERY
 static
 ss_system *
 readxyz(const char * filename)
@@ -580,25 +577,22 @@ calctex_cb(void * closure, double * pos, float elevation, double * spacing)
     return 0xffff0000;
   }
 }
-#endif // HAVE_LIBSCENERY
 
 void 
 Scenery::filenamesensor_cb(void * data, SoSensor * sensor)
 {
   Scenery * thisp = (Scenery*) data;
 
-  if (thisp->system) {
-#ifdef HAVE_LIBSCENERY
+  if ( sc_scenery_available() && thisp->system) {
     sc_ssglue_view_deallocate(thisp->system, thisp->viewid);
     sc_ssglue_system_close(thisp->system);
-#endif // HAVE_LIBSCENERY
   }
   thisp->viewid = -1;
   thisp->system = NULL;
 
   SbString s = thisp->filename.getValue();
   if (s.getLength()) {
-#ifdef HAVE_LIBSCENERY
+  if ( sc_scenery_available() ) {
 #if SS_IMPORT_XYZ
     if ( s.find(".xyz") == (s.getLength() - 4) ) {
       thisp->system = readxyz(s.getString());
@@ -626,58 +620,49 @@ Scenery::filenamesensor_cb(void * data, SoSensor * sensor)
       sc_ssglue_view_enable(thisp->system, thisp->viewid);
       //      fprintf(stderr,"system: %p, viewid: %d\n", thisp->system, thisp->viewid);
     }
-#endif // HAVE_LIBSCENERY
+    }
   }
 }
 
 void 
 Scenery::preFrame(void)
 {
-#ifdef HAVE_LIBSCENERY
-  if (this->system) {
+  if ( sc_scenery_available() && this->system ) {
     sc_ssglue_view_pre_frame(this->system, this->viewid);
     cc_hash_apply(this->texhash, hash_inc_unused, NULL);
   }
-#endif // HAVE_LIBSCENERY
 }
 
 int 
 Scenery::postFrame(void)
 {
-#ifdef HAVE_LIBSCENERY
-  if (this->system) {
+  if (sc_scenery_available() && this->system) {
     this->deleteUnusedTextures();
     return sc_ssglue_view_post_frame(this->system, this->viewid);
   }
-#endif // HAVE_LIBSCENERY
   return 0;
 }
 
 void 
 Scenery::setBlockRottger(const float c)
 {
-#ifdef HAVE_LIBSCENERY
-  if (this->system) {
+  if (sc_scenery_available() && this->system) {
     sc_ssglue_view_set_evaluate_rottger_parameters(this->system, this->viewid, 16.0f, c);
   }
-#endif // HAVE_LIBSCENERY
 }
 
 void 
 Scenery::setLoadRottger(const float c)
 {
-#ifdef HAVE_LIBSCENERY
-  if (this->system) {
+  if (sc_scenery_available() && this->system) {
     sc_ssglue_view_set_load_rottger_parameters(this->system, this->viewid, 16.0f, c);
   }
-#endif // HAVE_LIBSCENERY
 }
 
 void 
 Scenery::refreshTextures(const int id)
 {
-#ifdef HAVE_LIBSCENERY
-  if (this->system) {
+  if (sc_scenery_available() && this->system) {
     sc_ssglue_system_refresh_runtime_texture2d(this->system, id);
 
     this->tmplist.truncate(0);
@@ -695,14 +680,12 @@ Scenery::refreshTextures(const int id)
       }
     }
   }
-#endif // HAVE_LIBSCENERY
 }
 
 
 int 
 Scenery::gen_pre_cb(void * closure, ss_render_pre_cb_info * info)
 {
-#ifdef HAVE_LIBSCENERY
   Scenery * thisp = (Scenery*) closure; 
 
   RenderState & renderstate = thisp->renderstate;
@@ -713,9 +696,6 @@ Scenery::gen_pre_cb(void * closure, ss_render_pre_cb_info * info)
                                    &renderstate.elevdata,
                                    &renderstate.normaldata);
   return 1;
-#else // !HAVE_LIBSCENERY
-  return 0;
-#endif // !HAVE_LIBSCENERY
 }
 
 ////////////// render ///////////////////////////////////////////////////////////
@@ -767,7 +747,6 @@ void
 Scenery::undefrender_cb(void * closure, const int x, const int y, const int len, 
                            const unsigned int bitmask_org)
 {
-#ifdef HAVE_LIBSCENERY
   Scenery * thisp = (Scenery*) closure; 
 
   Scenery::RenderState * renderstate = &thisp->renderstate;
@@ -832,14 +811,12 @@ Scenery::undefrender_cb(void * closure, const int x, const int y, const int len,
       glEnd();
     }
   }
-#endif // HAVE_LIBSCENERY
 }
 
 void 
 Scenery::render_cb(void * closure, const int x, const int y,
                       const int len, const unsigned int bitmask)
 {
-#ifdef HAVE_LIBSCENERY
   Scenery * thisp = (Scenery*) closure;
   
   Scenery::RenderState * renderstate = &thisp->renderstate;
@@ -926,7 +903,6 @@ Scenery::render_cb(void * closure, const int x, const int y,
     glEnd();
 #undef ELEVATION
   }
-#endif // HAVE_LIBSCENERY
 }
 
 
@@ -945,7 +921,6 @@ void
 Scenery::undefgen_cb(void * closure, const int x, const int y, const int len, 
                      const unsigned int bitmask_org)
 {
-#ifdef HAVE_LIBSCENERY
   Scenery * thisp = (Scenery*) closure; 
 
   Scenery::RenderState * renderstate = &thisp->renderstate;
@@ -992,14 +967,12 @@ Scenery::undefgen_cb(void * closure, const int x, const int y, const int len,
       thisp->endShape();
     }
   }
-#endif // HAVE_LIBSCENERY
 }
 
 void 
 Scenery::gen_cb(void * closure, const int x, const int y,
                    const int len, const unsigned int bitmask)
 {
-#ifdef HAVE_LIBSCENERY
   Scenery * thisp = (Scenery*) closure;
 
   Scenery::RenderState * renderstate = &thisp->renderstate;
@@ -1031,7 +1004,6 @@ Scenery::gen_cb(void * closure, const int x, const int y,
   thisp->endShape();
   
 #undef ELEVATION
-#endif // HAVE_LIBSCENERY
 }
 
 SoGLImage *
